requirements.txt
appdirs==1.4.4
asn1crypto==1.5.1
beem==0.24.26
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-shell==2.1
contourpy==1.1.0
cycler==0.11.0
diff-match-patch==20230430
ecdsa==0.18.0
fonttools==4.40.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.25.0
packaging==23.1
Pillow==10.0.0
prettytable==3.8.0
pycryptodomex==3.18.0
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
scrypt==0.8.20
six==1.16.0
tabulate==0.9.0
urllib3==2.0.3
wcwidth==0.2.6
websocket-client==1.6.1
